Delete the legacy source-preparation path
The Snakemake lane is the only producer now, so the parallel recipe
system goes: 23 per-source Justfiles, the per-source downloader modules
(source_download, source_unzip, the five bespoke fetchers,
source_download_filelist), the unpublished tarball step, and the legacy
chain's e2e test. ext_for moves into source_prep, filelist_urls into
source_mirror (their only remaining consumers). source_catalog drops
the Justfile-sniffing datum invariant — every prepped source has a
sidecar (source_prep writes one unconditionally). just source/sources
delegate to the Snakefile; docs describe the metadata-driven model.
