+++ CONTRIBUTING.md
@@ -79,7 +79,7 @@ flowchart LR
 
 ### Layout
 
-- `sources/<id>/` - one dir per source: `metadata.json` (attribution), `file_list.txt` (URLs), `Justfile` (its fetch→DEM recipe). [sources/README.md](sources/README.md) is the catalog + how to add one.
+- `sources/<id>/` - one dir per source: `metadata.json` (attribution + prep knobs), `file_list.txt` (upstream URLs). [sources/README.md](sources/README.md) is the catalog + how to add one.
 - `pipelines/` - the Python engine (`uv` project) + `Justfile`. Stages: source → aggregation → downsampling → bundle.
 - `worker/` - Cloudflare Worker (TypeScript) that serves the unified tile endpoint from R2.
 - `index.js`, `index.html` - Vite/MapLibre viewer (repo root).
@@ -91,7 +91,7 @@ flowchart LR
 
 Four stages — `source → aggregation → downsampling → bundle` — all writing under `pipelines/store/` (gitignored), one subdir per stage (`store/source`, `store/aggregation`, `store/pmtiles`, `store/contour`, `store/bundle`, …). Incrementality is content-hash keys (`keys.py`): each store artifact is **content-addressed** — its key (inputs ‖ the pipeline modules that produce it ‖ the resolved config it read ‖ the toolchain) rides in its filename (`<stem>-<key12>.<ext>`), so freshness is "the named file exists" and a stage rebuilds a name that isn't present. Code and config changes rebuild exactly their dependents, and `FORCE_REBUILD=1` is an escape hatch, not a correctness requirement (see [docs/build.md](docs/build.md#the-incremental-model)).
 
-- **source** (`source_*.py`, per `sources/<id>/`): fetch → datum offset → normalize to a 4326 COG → `bounds.csv` → coverage polygon → tarball → a generated `catalog.json` (the machine-facing item the build reads for priority/max_zoom/land_clamp/offset + the recipe hash the tile keys consume; `metadata.json` stays the hand-edited input and fallback). Each source owns its recipe (`sources/<id>/Justfile`) composing the shared steps. Priority derives from `(maxzoom, id)` — GEBCO loses (smallest maxzoom), so regional sources win in overlap.
+- **source** (the Snakemake lane: repo-root `Snakefile` + `pipelines/source_*.py`): per-asset fetch → content-keyed staging (archives/formats detected from bytes) → datum → normalize to a COG → `bounds.csv` → coverage polygon → a generated `catalog.json` (the machine-facing item the build reads for priority/max_zoom/land_clamp/offset + the recipe hash the tile keys consume; `metadata.json` stays the hand-edited input and fallback, and carries each source's prep knobs). Priority derives from `(maxzoom, id)` — GEBCO loses (smallest maxzoom), so regional sources win in overlap.
 - **aggregation** (`aggregation_*.py`): `cover` slices the planet into source-aware aggregation tiles (one CSV each); `run` reprojects each source by priority into a merged Float32 DEM (Gaussian seam feather), slope-smooths it (`smooth.py`), encodes Terrarium raster tiles, and forks contours (`contour_run.py`) and soundings (`soundings_run.py`) off the same merged DEM — each fork independently keyed, so e.g. a contour-levels change re-runs the contour fork without rewriting terrain pmtiles. Sources flagged `land_clamp` (GEBCO/EMODnet — coarse, no land/water concept) get their negative land pixels clamped to 0 against the OSM land mask (`landmask.py`, prep once with `just landmask`) right after warp, so shoreline cells don't paint false water/contours/soundings over land. A `depare_run.py` fork buckets the same merged DEM into depth-area partitions (ENC DEPARE: water between charted isobaths, `drval1`/`drval2` depth bounds) — the vector twin of the raster depth shading — and folds two more kinds into the one `depare` layer off the same `gdal_contour -p` pass and the OSM land + inland-water masks: green drying foreshore (the `[0, DRYING_CAP]` bucket seaward of the land line but kept inside mapped inland water, negative `drval1`) and unknown-depth water (`nodata`, no `drval1`).
 - **downsampling**: 2×2-average overview pyramid below each source's native maxzoom; an overview's key hashes its children's keys, so child changes cascade up the pyramid by construction.
 - **bundle** (`bundle.py`): concat single-zoom pmtiles into `planet.pmtiles` + one overlay archive per populated grid cell + `manifest.json`; contours, soundings, and depare layers bundle into `vector.pmtiles` via tippecanoe; the per-source coverage footprints bundle standalone into `coverage.pmtiles` (`just coverage`). Bundles key off their members' keys and skip when nothing changed.
