--- sources/african_great_lakes/Justfile
@@ -1,15 +1,0 @@
-# African Great Lakes (GLWNB-2020) — Victoria, Albert, Edward, George. Run from pipelines/:  just ../sources/african_great_lakes/
-# One .7z on Harvard Dataverse; source_download_african_lakes extracts the 4 lake DEMs
-# via GDAL /vsi7z (no 7z binary). Values are water DEPTH (positive down, 0 at the lake
-# surface) → --negate, no offset. Per-lake CRS (Victoria Africa-LCC, Albert/George
-# UTM36N, Edward UTM35S) → mixed_crs: normalize WITHOUT --crs to keep each file's CRS;
-# aggregation reprojects each lake from its own frame. Lakes are hydraulically isolated,
-# so no ocean seam.
-[no-cd]
-default:
-    uv run python source_download_african_lakes.py african_great_lakes
-    uv run python source_datum.py african_great_lakes --negate
-    uv run python source_normalize.py african_great_lakes
-    uv run python source_bounds.py african_great_lakes
-    uv run python source_polygonize.py african_great_lakes 8
-    uv run python source_create_tarball.py african_great_lakes
