--- sources/ausbathytopo/Justfile
@@ -1,11 +1,0 @@
-# AusBathyTopo 250 m — national Australia. Run from pipelines/:  just ../sources/ausbathytopo/
-# One zip → one Float32 COG; MSL elevation (seafloor negative, like GEBCO) → no datum transform.
-# Already EPSG:4326; normalize re-asserts it, warp happens in aggregation. Native ~250 m → z9.
-[no-cd]
-default:
-    uv run python source_download.py ausbathytopo
-    uv run python source_unzip.py ausbathytopo
-    uv run python source_normalize.py ausbathytopo --crs EPSG:4326
-    uv run python source_bounds.py ausbathytopo
-    uv run python source_polygonize.py ausbathytopo 8
-    uv run python source_create_tarball.py ausbathytopo
