--- sources/batnas/Justfile
@@ -1,14 +1,0 @@
-# BATNAS Batimetri Nasional (Indonesia, ~180 m) — MSL. Run from pipelines/:  just ../sources/batnas/
-# Standard prepared source: fetches our R2 MIRROR of the raw BIG sheets (file_list points at
-# data.openwaters.io/.../mirror/batnas/) because BATNAS's official download is reCAPTCHA-login-
-# gated and not CI-fetchable. To refresh the mirror when BIG ships a new version, run
-# sources/batnas/harvest.py (needs a one-time login token) and re-upload to R2 — see that script.
-# Tiles have NO embedded CRS → normalize assigns EPSG:4326. MSL elevation (seafloor negative,
-# land positive, like GEBCO) → no datum transform.
-[no-cd]
-default:
-    uv run python source_download.py batnas
-    uv run python source_normalize.py batnas --crs EPSG:4326
-    uv run python source_bounds.py batnas
-    uv run python source_polygonize.py batnas 8
-    uv run python source_create_tarball.py batnas
