--- sources/bodensee/Justfile
@@ -1,14 +1,0 @@
-# Bodensee (Lake Constance) 3 m — IGKB Tiefenschärfe. Run from pipelines/:  just ../sources/bodensee/
-# One ESRI ASCII grid in a ~153 MB zip; source_download_swissbathy mosaics it to a tif.
-# Values are bed elevation in DHHN92 (orthometric, m above sea level); lake surface ~395.2 m
-# → --offset -395.2 makes the surface ~0, the bed negative; --clamp-positive drops the
-# surrounding-terrain fringe. EPSG:25832 (DHHN92-UTM32N) assigned in normalize. Isolated lake,
-# no ocean seam. (PANGAEA also ships an ETRS89 variant — ellipsoidal height — avoid: messier offset.)
-[no-cd]
-default:
-    uv run python source_download_swissbathy.py bodensee
-    uv run python source_datum.py bodensee --offset -395.2 --clamp-positive
-    uv run python source_normalize.py bodensee --crs EPSG:25832
-    uv run python source_bounds.py bodensee
-    uv run python source_polygonize.py bodensee 8
-    uv run python source_create_tarball.py bodensee
