--- sources/ddm/Justfile
@@ -1,10 +1,0 @@
-# Danmarks Dybdemodel — Danish waters. Run from pipelines/:  just ../sources/ddm/
-# Plain HTTP download; stored as positive-down depth → negate (source_datum).
-[no-cd]
-default:
-    uv run python source_download.py ddm
-    uv run python source_datum.py ddm --negate
-    uv run python source_normalize.py ddm --crs EPSG:3034
-    uv run python source_bounds.py ddm
-    uv run python source_polygonize.py ddm 8
-    uv run python source_create_tarball.py ddm
