--- sources/emodnet/Justfile
@@ -1,13 +1,0 @@
-# EMODnet bathymetry DTM 2024 — full European coverage (58 tiles, ~115 m native).
-# Each tile is a zipped GeoTIFF from the EMODnet download server (file_list.txt);
-# source_download fetches them, source_unzip extracts the .tif. No datum transform.
-# NOTE: ~7 GB zipped / tens of GB unzipped — heavier than a standard CI runner's
-# disk; stage it or use a larger runner.
-[no-cd]
-default:
-    uv run python source_download.py emodnet
-    uv run python source_unzip.py emodnet
-    uv run python source_normalize.py emodnet --crs EPSG:4326
-    uv run python source_bounds.py emodnet
-    uv run python source_polygonize.py emodnet 8
-    uv run python source_create_tarball.py emodnet
