--- sources/gsc_atlantic/Justfile
@@ -1,12 +1,0 @@
-# GSC Atlantic Bathymetric Compilation 100 m — Atlantic Canada. Run from pipelines/:  just ../sources/gsc_atlantic/
-# One ~210 MB GeoTIFF in a Lambert Conformal Conic projection (CRS embedded) → no --crs; the
-# warp to web mercator happens in aggregation. Continuous compiled DEM (not sparse like NONNA),
-# so the coverage polygon is one solid footprint. Topo-bathy → elevation (seafloor negative,
-# like GEBCO) → no datum transform. (Verify sign on first build; vertical datum unconfirmed.)
-[no-cd]
-default:
-    uv run python source_download.py gsc_atlantic
-    uv run python source_normalize.py gsc_atlantic
-    uv run python source_bounds.py gsc_atlantic
-    uv run python source_polygonize.py gsc_atlantic 8
-    uv run python source_create_tarball.py gsc_atlantic
