--- sources/infomar_10m/Justfile
@@ -1,12 +1,0 @@
-# INFOMAR 10 m inshore bathymetry — Irish waters. Run from pipelines/:  just ../sources/infomar_10m/
-# One zip → one Float32 GeoTIFF with NO embedded CRS, so normalize ASSIGNS EPSG:4326 (--crs).
-# Values are LAT-referenced elevation (seafloor negative, like GEBCO) → no datum transform.
-# The 25 m shelf grid is the sibling source sources/infomar_25m (cudem/cudem_third pattern).
-[no-cd]
-default:
-    uv run python source_download.py infomar_10m
-    uv run python source_unzip.py infomar_10m
-    uv run python source_normalize.py infomar_10m --crs EPSG:4326
-    uv run python source_bounds.py infomar_10m
-    uv run python source_polygonize.py infomar_10m 8
-    uv run python source_create_tarball.py infomar_10m
