--- sources/infomar_25m/Justfile
@@ -1,12 +1,0 @@
-# INFOMAR 25 m shelf bathymetry — Irish EEZ. Run from pipelines/:  just ../sources/infomar_25m/
-# One zip → one Float32 GeoTIFF with NO embedded CRS, so normalize ASSIGNS EPSG:4326 (--crs).
-# LAT-referenced elevation (seafloor negative, like GEBCO) → no datum transform. Sibling of the
-# 10 m sources/infomar_10m (cudem/cudem_third pattern): priority gives 10 m inshore, 25 m on shelf.
-[no-cd]
-default:
-    uv run python source_download.py infomar_25m
-    uv run python source_unzip.py infomar_25m
-    uv run python source_normalize.py infomar_25m --crs EPSG:4326
-    uv run python source_bounds.py infomar_25m
-    uv run python source_polygonize.py infomar_25m 8
-    uv run python source_create_tarball.py infomar_25m
