--- sources/lac_neuchatel/Justfile
@@ -1,14 +1,0 @@
-# Lac de Neuchâtel — swissBATHY3D, ~1 m. Run from pipelines/:  just ../sources/lac_neuchatel/
-# One esriasciigrid.zip of ~311 .asc tiles; source_download_swissbathy mosaics them to a tif.
-# Values are bed elevation in LN02 (m above sea level); surface 429.3 m → --offset -429.3
-# makes the surface ~0, the bed negative. EPSG:2056 (CH1903+ LV95) assigned in normalize.
-# Lake is hydraulically isolated, so no ocean seam. The zip carries a surrounding-terrain
-# fringe (cells above lake level), so --clamp-positive drops post-offset >0 cells to nodata.
-[no-cd]
-default:
-    uv run python source_download_swissbathy.py lac_neuchatel
-    uv run python source_datum.py lac_neuchatel --offset -429.3 --clamp-positive
-    uv run python source_normalize.py lac_neuchatel --crs EPSG:2056
-    uv run python source_bounds.py lac_neuchatel
-    uv run python source_polygonize.py lac_neuchatel 8
-    uv run python source_create_tarball.py lac_neuchatel
