--- sources/noaa_estuarine/Justfile
@@ -1,13 +1,0 @@
-# NOAA NOS Estuarine Bathymetric DEMs (30 m) — 69 US estuaries (Chesapeake, SF Bay,
-# Puget Sound, Long Island Sound, Galveston, Tampa, …). Run from pipelines/:  just ../sources/noaa_estuarine/
-# netCDF per estuary on the NGDC THREDDS server; each embeds its own NAD83/UTM zone →
-# mixed_crs (normalize runs WITHOUT --crs to keep the per-file CRS; aggregation reprojects
-# each from its own zone). Values are bed elevation on the local tidal datum (~MLLW),
-# already GEBCO convention (bed negative, intertidal positive) → no negate, no offset.
-[no-cd]
-default:
-    uv run python source_download_estuarine.py noaa_estuarine
-    uv run python source_normalize.py noaa_estuarine
-    uv run python source_bounds.py noaa_estuarine
-    uv run python source_polygonize.py noaa_estuarine 8
-    uv run python source_create_tarball.py noaa_estuarine
