--- sources/noaa_s102/Justfile
@@ -1,10 +1,0 @@
-# NOAA S-102 Bathymetric Surface, US navigable waters. Run from pipelines/:  just ../sources/noaa_s102/
-# Mirrored model: list the public bucket prefix, dedupe re-issued cells (newest LastModified
-# wins), and register every product as a relative objects/<key> row + a mirror manifest;
-# the bytes are copied into the store's object mirror, and aggregation range-reads them
-# from there — never from NOAA. Per-tile UTM zones + depth (positive-down): metadata sets
-# mixed_crs + band 1 + negate (depth->elevation at reproject, since mirrored sources skip
-# source_datum). Already MLLW — no datum offset. Sibling of BlueTopo but datum-clean.
-[no-cd]
-default:
-    uv run python source_mirror.py noaa_s102
