--- sources/swiobc/Justfile
@@ -1,11 +1,0 @@
-# swIOBC 250 m — SW Indian Ocean / E. Africa. Run from pipelines/:  just ../sources/swiobc/
-# One ~711 MB GeoTIFF, EPSG:4326, elevation (seafloor negative, like GEBCO) → no datum
-# transform. Fully populated (no nodata holes). normalize re-asserts the CRS; the warp to
-# web mercator happens in aggregation. Native ~250 m → z9 (omit max_zoom, let cover derive).
-[no-cd]
-default:
-    uv run python source_download.py swiobc
-    uv run python source_normalize.py swiobc --crs EPSG:4326
-    uv run python source_bounds.py swiobc
-    uv run python source_polygonize.py swiobc 8
-    uv run python source_create_tarball.py swiobc
