--- sources/uk_surfzone/Justfile
@@ -1,19 +1,0 @@
-# SurfZone DEM 2019 (Environment Agency, England intertidal/surf zone, 2 m) — OGL v3.0.
-# Run from pipelines/:  just ../sources/uk_surfzone/
-# 1301 per-tile ZIPs fetched over plain HTTPS from the Defra survey tile API (no login; the
-# public ?subscription-key=dspui is baked into file_list.txt). Each ZIP holds one
-# SurfZone_DEM_2019_<tile>_2m.tif — source_download saves the extensionless URL as .zip
-# (magic-byte sniff), source_unzip flattens the tif out. CRS is embedded (EPSG:27700) but we
-# assign it anyway to cover any tile that ships without one. Vertical datum is Ordnance Datum
-# Newlyn (~MSL); elevation, seafloor negative (like GEBCO) → no datum transform.
-# Heavy prep: ~4 GB of zips → ~30 GB of extracted Float32 tifs before they renormalize to ZSTD
-# COGs (one-time, on a build runner). Regenerate file_list.txt with harvest.py if the dspui key
-# rotates or the EA ships a new vintage.
-[no-cd]
-default:
-    uv run python source_download.py uk_surfzone
-    uv run python source_unzip.py uk_surfzone
-    uv run python source_normalize.py uk_surfzone --crs EPSG:27700
-    uv run python source_bounds.py uk_surfzone
-    uv run python source_polygonize.py uk_surfzone 8
-    uv run python source_create_tarball.py uk_surfzone
