--- sources/vaklodingen/Justfile
@@ -1,11 +1,0 @@
-# Vaklodingen 20 m — Dutch coastal & inland waters. Run from pipelines/:  just ../sources/vaklodingen/
-# Plain HTTP download of one GeoTIFF; values are NAP bed elevation (seafloor negative,
-# like GEBCO) → no datum transform. EPSG:28992 is embedded; normalize re-asserts it and
-# the warp to web mercator happens later in aggregation.
-[no-cd]
-default:
-    uv run python source_download.py vaklodingen
-    uv run python source_normalize.py vaklodingen --crs EPSG:28992
-    uv run python source_bounds.py vaklodingen
-    uv run python source_polygonize.py vaklodingen 8
-    uv run python source_create_tarball.py vaklodingen
